--- src/main/kotlin/io/github/filippovissani/kotlin_distributed_dfrp/core/extensions/ExportTreeManipulation.kt
@@ -1,7 +1,0 @@
-package io.github.filippovissani.kotlin_distributed_dfrp.core.extensions
-
-import io.github.filippovissani.kotlin_distributed_dfrp.core.ExportTree
-
-fun <T, R> ExportTree<T>.map(transform: (T) -> R): ExportTree<R> {
-    return ExportTree(transform(root), this.children)
-}
