--- src/main/kotlin/io/github/filippovissani/kotlin_distributed_dfrp/samples/Gradient.kt
@@ -1,4 +1,0 @@
-package io.github.filippovissani.kotlin_distributed_dfrp.samples
-
-class Gradient {
-}
